# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dzaïr Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Wilaya {
  id: number;
  code: number;
  name: string;
  nameAr: string;
  region: string;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  icon: string;
  listingCount: number;
}

export interface Store {
  id: number;
  name: string;
  nameAr: string;
  description: string;
  wilayaId: number;
  wilayaName: string;
  categoryId: number;
  categoryName: string;
  phone: string;
  address: string;
  logoUrl?: string;
  coverUrl?: string;
  rating: number;
  reviewCount: number;
  listingCount: number;
  isVerified: boolean;
  isFeatured: boolean;
  createdAt: string;
}

export interface Listing {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  price: number;
  currency: string;
  imageUrl?: string;
  storeId: number;
  storeName: string;
  categoryId: number;
  categoryName: string;
  wilayaId: number;
  wilayaName: string;
  condition: string;
  isFeatured: boolean;
  isNegotiable: boolean;
  viewCount: number;
  createdAt: string;
}

export type StoreDetail = Store & {
  email?: string;
  website?: string;
  openingHours?: string;
  listings: Listing[];
};

export type ListingDetail = Listing & {
  images: string[];
  store: Store;
  relatedListings: Listing[];
};

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  totalPages: number;
}

export interface SearchResults {
  listings: Listing[];
  stores: Store[];
  total: number;
}

export interface MarketplaceStats {
  totalListings: number;
  totalStores: number;
  totalWilayas: number;
  totalCategories: number;
}

export type GetListingsParams = {
  categoryId?: number;
  wilayaId?: number;
  storeId?: number;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  featured?: boolean;
  page?: number;
  limit?: number;
};

export type GetStoresParams = {
  wilayaId?: number;
  categoryId?: number;
  search?: string;
  featured?: boolean;
};

export type SearchAllParams = {
  q: string;
};