# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dzaïr Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  price: number;
  currency: string;
  imageUrl?: string;
  storeId: number;
  storeName: string;
  categoryId: number;
  categoryName: string;
  wilayaId: number;
  wilayaName: string;
  condition: string;
  isFeatured: boolean;
  isNegotiable: boolean;
  viewCount: number;
  createdAt: string;
}