# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dzaïr Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from "./listing";
import type { Store } from "./store";

export interface SearchResults {
  listings: Listing[];
  stores: Store[];
  total: number;
}