# lib/api-zod/src/generated/types/wilaya.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dzaïr Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Wilaya {
  id: number;
  code: number;
  name: string;
  nameAr: string;
  region: string;
}